Move the project types to the lib
In order to avoid deep relative imports, move the types file to the library
