+++ apps/web/src/routes/[ownerSlug]/[projectSlug]/+page.ts
@@ -1,5 +1,5 @@
+import type { ProjectParameters } from '$lib/project/types';
 import type { PageLoad } from './$types';
-import type { ProjectParameters } from '../types';
 
 // eslint-disable-next-line func-style
 export const load: PageLoad<ProjectParameters> = async ({ params }) => {
