+++ apps/web/src/routes/[ownerSlug]/[projectSlug]/reviews/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { ProjectParameters } from '../../types';
+	import type { ProjectParameters } from '$lib/project/types';
 
 	interface Props {
 		data: ProjectParameters;
